+++ webroot/js/browserRedirect.js
@@ -1,0 +1,11 @@
+import { exec } from './kernelsu.js'
+
+document.addEventListener('click', async (event) => {
+  const getLink = event.target.getAttribute('credit-link')
+  if (!getLink || typeof getLink !== 'string') return;
+
+  const ptrace64Cmd = await exec(`am start -a android.intent.action.VIEW -d https://${getLink}`).catch(() => {
+    return window.open(`https://${getLink}`, "_blank", 'toolbar=0,location=0,menubar=0')
+  })
+  if (ptrace64Cmd.errno !== 0) return window.open(`https://${getLink}`, "_blank", 'toolbar=0,location=0,menubar=0')
+}, false)
