+++ webroot/js/daemonActions.js
