+++ webroot/js/errorScreen.js
@@ -1,0 +1,6 @@
+const button = document.getElementById('esc-copy-button')
+const log = document.getElementById('esc-log')
+
+button.addEventListener('click', () => {
+  navigator.clipboard.writeText(log.innerHTML)
+})
